refactor(process-cooling): move confirm-action component to shared
