--- src/app/process-cooling-assessment/confirm-action/confirm-action.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ConfirmActionComponent } from './confirm-action.component';
-
-describe('ConfirmActionComponent', () => {
-  let component: ConfirmActionComponent;
-  let fixture: ComponentFixture<ConfirmActionComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ConfirmActionComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(ConfirmActionComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
