+++ src/app/process-cooling-assessment/results-panel/inventory-table/inventory-table.component.ts
@@ -3,7 +3,7 @@ import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { ChillerInventoryItem } from '../../../shared/models/process-cooling-assessment';
 import { ChillerInventoryBaseComponent } from './chiller-inventory-base.component';
 import { ModalDialogService } from '../../../shared/modal-dialog.service';
-import { ConfirmActionComponent, ConfirmActionData } from '../../confirm-action/confirm-action.component';
+import { ConfirmActionComponent, ConfirmActionData } from '../../../shared/confirm-action/confirm-action.component';
 import { FilterChillerInventoryParams } from '../../pipes/filter-chiller-inventory.pipe';
 
 @Component({
